Move AdminActionError to errors folder
