+++ src/scripts/utils/errors/AdminActionError.js
